Add global variables for calculation
Currently just committing with every number on the list
because I haven't been sold/convinced on the commit early and often
thing. I can conceptually get it but is there a too early a too often
and am I hitting those goals? A whole commit for 3 declared variables
feels wrong but maybe it's a standard.